fix test output for parent tests
We want the test output for parent tests to reflect the status
of the subtests (i.e. the messages shouldn't contain "FAIL"
when a failed subtest was skipped), but there's an awkwardness
there because the parent test outputs "FAIL" before any of the
status lines have been emitted for its subtests.

We solve that by buffering output for a parent test until
its status is known.